refactor and added tons of unused code
