+++ .gitignore
@@ -11,5 +11,6 @@ node_modules/
 *.png
 *.jpg
 
+# data directory stuff
 *.traineddata
 data/*.png
